View de horario foi transformada em uma tag
