+++ escola/models.py
@@ -1,3 +1,5 @@
+import logging
+
 from django.contrib.auth.models import User, Group
 from django.db import models
 from django_prometheus.models import ExportModelOperationsMixin
@@ -6,9 +8,12 @@ from guardian.shortcuts import assign_perm
 
 
 # Create your models here.
+import escola
 
+logger = logging.getLogger(__name__)
 
 class Profile(models.Model, ExportModelOperationsMixin('Profiles')):
+    """Perfil basico, todos os usuarios do sistema tem um."""
     user = models.OneToOneField(User, related_name='profile_escola', on_delete=models.CASCADE)
     is_aluno = models.BooleanField('student status', default=False)
     is_professor = models.BooleanField('teacher status', default=False)
@@ -36,6 +41,7 @@ class Profile(models.Model, ExportModelOperationsMixin('Profiles')):
 
 
 class Turma(models.Model, ExportModelOperationsMixin('Turma')):
+    """ Uma turma, conjunto de alunos, materias, tarefas, tambem possui um horario"""
     numero = models.IntegerField()
     ano = models.IntegerField()
     lider = models.ForeignKey(Group, on_delete=models.DO_NOTHING, null=True, blank=True, related_name='turma_lider')
@@ -95,60 +101,17 @@ class Turma(models.Model, ExportModelOperationsMixin('Turma')):
             assign_perm('escola.can_add_materia', self.regente, obj=self)
             assign_perm('escola.can_add_tarefa', self.regente, obj=self)
             return self.regente
+
+    def get_or_create_horario(self):
+        try:
+            return self.horario
+        except escola.models.Turma.horario.RelatedObjectDoesNotExist:
+            h = Horario(turma=self)
+            h.save()
+            return h
+
     def __str__(self):
         return f"Turma {self.numero}"
-
-    def is_lider(self, user: User):
-        lider = self.get_lider()
-        if lider is None:
-            return False
-        elif lider == user:
-            return True
-        else:
-            return False
-
-    def is_regente(self, user: User):
-        regente = self.get_regente()
-        if regente is None:
-            return False
-        elif regente == user:
-            return True
-        else:
-            return False
-
-    def is_user_especial(self, user: User):
-        cargos = self.get_cargo_especial_list()
-        if len(cargos) == 0:
-            return False
-        else:
-            i = 0
-            while i < len(cargos) and not cargos[i].ocupante == user:
-                i = i + 1
-            if i < len(cargos):
-                return False
-            else:
-                return True
-
-    def get_lider_list(self):
-        return self.cargoturma_set.filter(cod_especial=1)
-
-    def get_lider(self):
-        if self.get_lider_list().exists():
-            return self.get_lider_list()[0].ocupante
-        else:
-            return None
-
-    def get_regente_list(self):
-        return self.cargoturma_set.filter(cod_especial=5)
-
-    def get_regente(self):
-        if self.get_regente_list().exists():
-            return self.get_regente_list()[0].ocupante
-        else:
-            return None
-
-    def get_cargo_especial_list(self):
-        return self.cargoturma_set.all()
 
 
 class SeguidorManager(models.Model, ExportModelOperationsMixin('SeguidorManager')):
@@ -309,51 +272,105 @@ class Notificacao(models.Model, ExportModelOperationsMixin('Alerta')):
     msg = models.TextField()
     link = models.URLField(blank=True, null=True)
 
-# class Conteudo(models.Model):
-#     materias = models.ManyToManyField(MateriaDaTurma)
-#     nome = models.CharField(max_length=50)
-#     serie = models.IntegerField() # 1 => 1 fundamente, 10=> 1º Medio, 20, 30
-#     parent = models.ForeignKey('self', null=True, blank=True, on_delete=models.CASCADE)
-#
-# class CaixaBase(models.Model, ExportModelOperationsMixin('CaixaBase')):
-#     objects = InheritanceManager()
-#     nome = 'Classe Base'
-#     pode_usar = False
-#     min_w = 0
-#     min_h = 0
-#     max_h = 4
-#     max_w = 3
-#
-#     config_url = ''
-#
-#     pos_x = models.IntegerField()
-#     pos_y = models.IntegerField()
-#
-#     def return_html(self):
-#         return 'Não implementado - Usando metodo da CaixaBase'
-#
-#
-# # class ConteudosCaixa(CaixaBase):
-# #     min_h = 2
-# #     min_w = 1
-# #     max_w = 2
-# #     pode_usar = True
-# #     nome = 'Box de Conteudos'
-# #
-# #     conteudos = models.ManyToManyField(Conteudo)
-# #
-# #     def return_html(self):
-# #         return 'Ainda não implementado - Listará Conteudos'
-#
-# class TestCaixa(CaixaBase):
-#     min_h = 2
-#     min_w = 1
-#     max_w = 2
-#     pode_usar = True
-#     nome = 'Box de Teste - Se você encontrar está entre em contato'
-#
-#     text = models.TextField()
-#
-#     def return_html(self):
-#         return 'Box de Teste - Se você encontrar está entre em contato </br>' \
-#                ''
+
+class Horario(models.Model, ExportModelOperationsMixin('Horario')):
+    turma = models.OneToOneField(Turma, related_name='horario', on_delete=models.CASCADE)
+
+    def get_turno_aula_or_create(self, dia, turno_a):
+        turno = TurnoAula.objects.filter(diaDaSemana=dia, turno=turno_a, turma=self.turma)
+        if turno:
+            return turno[0]
+        else:
+            turno = TurnoAula(turma=self.turma, horario=self, diaDaSemana=dia, turno=turno_a)
+            turno.save()
+            return turno
+
+    def get_periodo_or_create(self, dia, turno: int, num):
+        turno_aula = self.get_turno_ausla_or_create(dia, Turno.get_turno_by_cod(turno))
+        per = turno_aula.periodo_set.filter(num=num)
+        if per:
+            return per[0]
+        else:
+            per = Periodo(turnoAula=turno_aula, num=num)
+            per.save()
+            return per
+
+    def get_horario(self):
+        logger.debug('horario:get_horario()')
+        turnos = Turno.objects.all().order_by('cod')
+        logger.info('Puxou %s turno(s) do banco de dados.', len(turnos))
+        DIAS_DA_SEMANA_N = range(1, 8)
+        ta = {}
+        logger.info('Preparando para entrar no loop de turnos...')
+        for turno in turnos:
+            logger.debug('Turno id:%s', turno.pk)
+            # Passa todos os dias da semana
+            for dia in DIAS_DA_SEMANA_N:
+                # Pega Turnos de Aula em que o turno, dia e horario seja certo;
+                a = TurnoAula.objects.filter(turno=turno, diaDaSemana=dia, horario=self)
+                # Se houver um turno
+                if len(a) > 0:
+                    # Se esse dia não estiver na lista de horario que sai adiciona
+                    if dia not in ta:
+                        ta[dia] = dict()
+                    # Salva o turno aula na saida
+                    ta[dia][turno.cod] = a[0]
+        return ta
+
+
+class Turno(models.Model, ExportModelOperationsMixin('Turno')):
+    nome = models.CharField(max_length=30)
+    cod = models.PositiveSmallIntegerField()
+    horaInicio = models.TimeField()
+    # 1
+    s1 = models.TimeField()
+    # 2
+    s2 = models.TimeField()
+    # 3
+    s3 = models.TimeField()
+    # intervalo
+    s4 = models.TimeField()
+    # 4
+    s5 = models.TimeField()
+    # 5
+    horaFim = models.TimeField()
+
+    def __str__(self):
+        return self.nome
+
+    def get_turno_by_cod(cod):
+        return Turno.objects.filter(cod=cod)[0]
+
+
+class TurnoAula(models.Model, ExportModelOperationsMixin('TurnoAula')):
+    turma = models.ForeignKey(Turma, on_delete=models.CASCADE)
+    horario = models.ForeignKey(Horario, on_delete=models.CASCADE)
+    DIAS_DA_SEMANA = (
+        (1, 'domingo'),
+        (2, 'segunda-feira'),
+        (3, 'terca-feira'),
+        (4, 'quarta-feira'),
+        (5, 'quinta-feira'),
+        (6, 'sexta-feira'),
+        (7, 'sabado'),
+    )
+    diaDaSemana = models.PositiveSmallIntegerField(choices=DIAS_DA_SEMANA)
+    turno = models.ForeignKey(Turno, on_delete=models.CASCADE)
+
+
+class Periodo(models.Model, ExportModelOperationsMixin('Periodo')):
+    num = models.PositiveSmallIntegerField()
+    turnoAula = models.ForeignKey(TurnoAula, on_delete=models.CASCADE)
+    materia = models.ForeignKey(MateriaDaTurma, on_delete=models.CASCADE, null=True, blank=True)
+
+    @property
+    def dia(self):
+        return self.turnoAula.diaDaSemana
+
+    @property
+    def turma(self):
+        return self.turnoAula.turno.turma
+
+    @property
+    def turno_cod(self):
+        return self.turnoAula.turno.cod
