--- escola/models_horario.py
@@ -1,85 +1,0 @@
-from django.db import models
-from django_prometheus.models import ExportModelOperationsMixin
-
-from escola.models import Turma, MateriaDaTurma
-
-
-class Horario(models.Model, ExportModelOperationsMixin('Horario')):
-    turma = models.OneToOneField(Turma, related_name='horario', on_delete=models.CASCADE)
-
-    def get_turno_aula_or_create(self, dia, turno_a):
-        turno = TurnoAula.objects.filter(diaDaSemana=dia, turno=turno_a, turma=self.turma)
-        if turno:
-            return turno[0]
-        else:
-            turno = TurnoAula(turma=self.turma, horario=self, diaDaSemana=dia, turno=turno_a)
-            turno.save()
-            return turno
-
-    def get_periodo_or_create(self, dia, turno: int, num):
-        turno_aula = self.get_turno_ausla_or_create(dia, Turno.get_turno_by_cod(turno))
-        per = turno_aula.periodo_set.filter(num=num)
-        if per:
-            return per[0]
-        else:
-            per = Periodo(turnoAula=turno_aula, num=num)
-            per.save()
-            return per
-
-
-class Turno(models.Model, ExportModelOperationsMixin('Turno')):
-    nome = models.CharField(max_length=30)
-    cod = models.PositiveSmallIntegerField()
-    horaInicio = models.TimeField()
-    # 1
-    s1 = models.TimeField()
-    # 2
-    s2 = models.TimeField()
-    # 3
-    s3 = models.TimeField()
-    # intervalo
-    s4 = models.TimeField()
-    # 4
-    s5 = models.TimeField()
-    # 5
-    horaFim = models.TimeField()
-
-    def __str__(self):
-        return self.nome
-
-    def get_turno_by_cod(cod):
-        return Turno.objects.filter(cod=cod)[0]
-
-
-class TurnoAula(models.Model, ExportModelOperationsMixin('TurnoAula')):
-    turma = models.ForeignKey(Turma, on_delete=models.CASCADE)
-    horario = models.ForeignKey(Horario, on_delete=models.CASCADE)
-    DIAS_DA_SEMANA = (
-        (1, 'domingo'),
-        (2, 'segunda-feira'),
-        (3, 'terca-feira'),
-        (4, 'quarta-feira'),
-        (5, 'quinta-feira'),
-        (6, 'sexta-feira'),
-        (7, 'sabado'),
-    )
-    diaDaSemana = models.PositiveSmallIntegerField(choices=DIAS_DA_SEMANA)
-    turno = models.ForeignKey(Turno, on_delete=models.CASCADE)
-
-
-class Periodo(models.Model, ExportModelOperationsMixin('Periodo')):
-    num = models.PositiveSmallIntegerField()
-    turnoAula = models.ForeignKey(TurnoAula, on_delete=models.CASCADE)
-    materia = models.ForeignKey(MateriaDaTurma, on_delete=models.CASCADE, null=True, blank=True)
-
-    @property
-    def dia(self):
-        return self.turnoAula.diaDaSemana
-
-    @property
-    def turma(self):
-        return self.turnoAula.turno.turma
-
-    @property
-    def turno_cod(self):
-        return self.turnoAula.turno.cod
