+++ escola/templates/escola/home.html
@@ -1,6 +1,6 @@
 {% extends "escola/base.html" %}
 <!--FIXME Implementar permissões melhoradas-->
-
+{%load dict_key %}
 {% block title %} Home {% endblock %}
 
 
@@ -25,42 +25,7 @@
 {% if request.user.profile_escola.is_aluno%}
 <article>
     <h2>Horario:</h2>
-<table>
-  <tr>
-      <th></th>
-      {%for d in DIAS_DA_SEMANA%}
-         <th>
-          {{d}}
-          </th>
-      {%endfor%}
-
-  </tr>
-   {% for turno in turnos %}
-    <tr>
-        <th>{{turno.nome}}</th>
-        {%for dia in DIAS_DA_SEMANA_N%}
-        <th>
-
-
-           {%dia_teste turno dia ta as verdade%}
-            {%if verdade%}
-                {%for p in range%}
-                    {%get_p turno dia p ta%} <br />
-                {%endfor%}
-                {%if perms.escola.editar_horario%}
-                <a href="{%url 'escola:alterar-horario' turma_pk turno.cod dia%}">EDITAR</a>
-                {%endif%}
-            {%else%}
-                Não há um horario cadastrado para esse horario.
-                {%if perms.escola.editar_horario%}
-                <a href="{%url 'escola:alterar-horario' turma_pk turno.cod dia%}">EDITAR</a>
-                {%endif%}
-            {%endif%}
-        </th>
-        {%endfor%}
-    </tr>
-    {%endfor%}
-</table>
+{% mostra_horario_proprio user %}
 </article>
 <article>
     <h2>Proximas tarefas</h2>
