+++ escola/templates/escola/horario/horario_include.html
@@ -1,0 +1,39 @@
+{%load dict_key %}
+{% load guardian_tags %}
+{% get_obj_perms user for turma as "turma_perm" %}
+<table>
+  <tr>
+      <th></th>
+      {%for d in DIAS_DA_SEMANA%}
+         <th>
+          {{d}}
+          </th>
+      {%endfor%}
+
+  </tr>
+   {% for turno in turnos %}
+    <tr>
+        <th>{{turno.nome}}</th>
+        {%for dia in DIAS_DA_SEMANA_N%}
+        <th>
+
+
+           {%dia_teste turno dia ta as verdade%}
+            {%if verdade%}
+                {%for p in range%}
+                    {%get_p turno dia p ta%} <br />
+                {%endfor%}
+                {%if 'editar_horario' in turma_perm%}
+                <a href="{%url 'escola:alterar-horario' turma_pk turno.cod dia%}">EDITAR</a>
+                {%endif%}
+            {%else%}
+                Não há um horario cadastrado para esse horario.
+                {%if 'editar_horario' in turma_perm%}
+                <a href="{%url 'escola:alterar-horario' turma_pk turno.cod dia%}">EDITAR</a>
+                {%endif%}
+            {%endif%}
+        </th>
+        {%endfor%}
+    </tr>
+    {%endfor%}
+</table>
