+++ escola/templates/escola/horario/mostraHorario.html
@@ -3,41 +3,43 @@
 {%load dict_key %}
 {% block conteudo %}
 {# turnos , DIAS_DA_SEMANA , DIAS_DA_SEMANA_N, ta[dia][turno.cod]#}
-<table>
-  <tr>
-      <th></th>
-      {%for d in DIAS_DA_SEMANA%}
-         <th>
-          {{d}}
-          </th>
-      {%endfor%}
-      
-  </tr>
-   {% for turno in turnos %}
-    <tr>
-        <th>{{turno.nome}}</th>
-        {%for dia in DIAS_DA_SEMANA_N%}
-        <th>
-           
-           
-           {%dia_teste turno dia ta as verdade%}
-            {%if verdade%}
-                {%for p in range%}
-                    {%get_p turno dia p ta%} <br />
-                {%endfor%}
-                {%if perms.escola.editar_horario%}
-                <a href="{%url 'escola:alterar-horario' turma_pk turno.cod dia%}">EDITAR</a>
-                {%endif%}
-            {%else%}
-                Não há um horario cadastrado para esse horario.
-                {%if perms.escola.editar_horario%}
-                <a href="{%url 'escola:alterar-horario' turma_pk turno.cod dia%}">EDITAR</a>
-                {%endif%}
-            {%endif%}
-        </th>
-        {%endfor%}
-    </tr>
-    {%endfor%}
-</table>
+<!--<table>-->
+  <!--<tr>-->
+      <!--<th></th>-->
+      <!--{%for d in DIAS_DA_SEMANA%}-->
+         <!--<th>-->
+          <!--{{d}}-->
+          <!--</th>-->
+      <!--{%endfor%}-->
+      <!---->
+  <!--</tr>-->
+   <!--{% for turno in turnos %}-->
+    <!--<tr>-->
+        <!--<th>{{turno.nome}}</th>-->
+        <!--{%for dia in DIAS_DA_SEMANA_N%}-->
+        <!--<th>-->
+           <!---->
+           <!---->
+           <!--{%dia_teste turno dia ta as verdade%}-->
+            <!--{%if verdade%}-->
+                <!--{%for p in range%}-->
+                    <!--{%get_p turno dia p ta%} <br />-->
+                <!--{%endfor%}-->
+                <!--{%if perms.escola.editar_horario%}-->
+                <!--<a href="{%url 'escola:alterar-horario' turma_pk turno.cod dia%}">EDITAR</a>-->
+                <!--{%endif%}-->
+            <!--{%else%}-->
+                <!--Não há um horario cadastrado para esse horario.-->
+                <!--{%if perms.escola.editar_horario%}-->
+                <!--<a href="{%url 'escola:alterar-horario' turma_pk turno.cod dia%}">EDITAR</a>-->
+                <!--{%endif%}-->
+            <!--{%endif%}-->
+        <!--</th>-->
+        <!--{%endfor%}-->
+    <!--</tr>-->
+    <!--{%endfor%}-->
+<!--</table>-->
+
+{%mostra_horario turma user%}
 
 {%endblock%}
